Rename: Changed app name server to haleygg for create new app while avoid naming conflicts.
