+++ haleygg/managers.py
