+++ haleygg/urls.py
@@ -2,11 +2,11 @@ from django.urls import include
 from django.urls import path
 from rest_framework.routers import DefaultRouter
 
-from server.views import MatchSummaryView
-from server.views import MatchViewSet
-from server.views import MapViewSet
-from server.views import LeagueViewSet
-from server.views import ProfileViewSet
+from haleygg.views import MatchSummaryView
+from haleygg.views import MatchViewSet
+from haleygg.views import MapViewSet
+from haleygg.views import LeagueViewSet
+from haleygg.views import ProfileViewSet
 
 
 router = DefaultRouter()
