+++ haleygg_server/settings/local.py
@@ -7,7 +7,7 @@ ALLOWED_HOSTS += ["127.0.0.1"]
 
 INSTALLED_APPS += [
     "debug_toolbar",
-    "server",
+    "haleygg",
 ]
 
 INTERNAL_IPS = [
