+++ haleygg_server/urls.py
@@ -19,5 +19,5 @@ from django.urls import path, include
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("__debug__/", include("debug_toolbar.urls")),
-    path("api/", include("server.urls")),
+    path("api/", include("haleygg.urls")),
 ]
